requirements.txt
numpy==1.23.2
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==9.1.1
scipy==1.8.1
skimage
tensorboardX==2.6
torchgeometry==0.1.2
tqdm==4.64.0